Alterando logica de autenticacao

+++ src/controllers/UserController.js
@@ -1,18 +1,24 @@
+const bcrypt = require("bcryptjs");
+
 const User = require("../models/Usuario");
 const Company = require("../models/Empresa");
 
 module.exports = {
   //### Cadastra um novo usuário
   async create(req, res) {
-    const _id = req.companyId;
-
-    const { email } = req.body;
+    const _id = req.userPayload.empresa;
+    const { nome } = req.body;
 
     try {
       if (!(await Company.findById({ _id })))
         return res.status(404).send({ error: "Empresa não encontrada" });
 
-      const user = await User.findOne({ email, id_empresa: _id });
+      if (!req.userPayload.admin)
+        return res.status(401).send({
+          Negado: "Usuário não possui permissão para realizar essa operação",
+        });
+
+      const user = await User.findOne({ nome, id_empresa: _id });
 
       if (user) return res.status(400).send({ error: "Usuário já cadastrado" });
 
@@ -29,9 +35,14 @@ module.exports = {
 
   //### Lista os usuários de uma empresa
   async list(req, res) {
-    const id_empresa = req.companyId;
+    const id_empresa = req.userPayload.empresa;
 
     try {
+      if (!req.userPayload.admin)
+        return res.status(401).send({
+          Negado: "Usuário não possui permissão para realizar essa operação",
+        });
+
       const users = await User.find({ id_empresa });
 
       const count = users.length;
@@ -60,18 +71,26 @@ module.exports = {
 
   //### ALtera informações de um usuário
   async update(req, res) {
-    const _id = req.userId;
-
     try {
-      const user = await User.findOne({ _id });
+      var hash;
+      const { senha } = req.body;
+
+      const user = await User.findById(req.userPayload.id).select("+senha");
+
+      if (senha) {
+        hash = await bcrypt.hash(senha, 10);
+      } else {
+        hash = user.senha;
+      }
 
       if (!user)
         return res.status(400).send({ error: "Usuário não encontrado" });
 
       const updateUser = await User.findByIdAndUpdate(
-        _id,
+        req.userPayload.id,
         {
           ...req.body,
+          senha: hash,
         },
         { new: true }
       );
@@ -79,16 +98,21 @@ module.exports = {
       return res.status(200).send({ updateUser });
     } catch (error) {
       console.log(error);
-      return res.status(400).send({ error });
+      return res.status(400).send({ error: error.message });
     }
   },
 
   //### Deleta um usuário
   async delete(req, res) {
     const _id = req.params.userId;
-    const id_empresa = req.companyId;
+    const id_empresa = req.userPayload.empresa;
 
     try {
+      if (!req.userPayload.admin)
+        return res.status(401).send({
+          Negado: "Usuário não possui permissão para realizar essa operação",
+        });
+
       const user = await User.findOneAndRemove({ _id, id_empresa });
 
       if (!user) return res.send({ error: "Usuário não encontrado" });
